add base chat assistant client
