+++ sdk/azure-ai-assistant/azure/ai/assistant/management/chat_assistant_client.py
@@ -4,17 +4,16 @@
 from azure.ai.assistant.management.assistant_config import AssistantConfig
 from azure.ai.assistant.management.assistant_config_manager import AssistantConfigManager
 from azure.ai.assistant.management.assistant_client_callbacks import AssistantClientCallbacks
-from azure.ai.assistant.management.base_assistant_client import BaseAssistantClient
+from azure.ai.assistant.management.base_chat_assistant_client import BaseChatAssistantClient
 from azure.ai.assistant.management.conversation_thread_client import ConversationThreadClient
 from azure.ai.assistant.management.exceptions import EngineError, InvalidJSONError
 from azure.ai.assistant.management.logger_module import logger
 from typing import Optional
 from datetime import datetime
-import json, uuid
-import copy
+import json, uuid, yaml
 
 
-class ChatAssistantClient(BaseAssistantClient):
+class ChatAssistantClient(BaseChatAssistantClient):
     """
     A class that manages an chat assistant client.
 
@@ -35,8 +34,6 @@ class ChatAssistantClient(BaseAssistantClient):
             timeout: Optional[float] = None
     ) -> None:
         super().__init__(config_json, callbacks)
-        self._tools = None
-        self._messages = []
         self._init_chat_assistant_client(self._config_data, is_create, timeout=timeout)
 
     @classmethod
@@ -67,6 +64,33 @@ class ChatAssistantClient(BaseAssistantClient):
             raise InvalidJSONError(f"Invalid JSON format: {e}")
 
     @classmethod
+    def from_yaml(
+        cls,
+        config_yaml: str,
+        callbacks: Optional[AssistantClientCallbacks] = None,
+        timeout: Optional[float] = None
+    ) -> "ChatAssistantClient":
+        """
+        Creates an ChatAssistantClient instance from YAML configuration data.
+
+        :param config_yaml: YAML string containing the configuration for the chat assistant.
+        :type config_yaml: str
+        :param callbacks: Optional callbacks for the chat assistant client.
+        :type callbacks: Optional[AssistantClientCallbacks]
+        :param timeout: Optional timeout for HTTP requests.
+        :type timeout: Optional[float]
+        :return: An instance of ChatAssistantClient.
+        :rtype: ChatAssistantClient
+        """
+        try:
+            config_data = yaml.safe_load(config_yaml)
+            config_json = json.dumps(config_data)
+            return cls.from_json(config_json, callbacks, timeout)
+        except yaml.YAMLError as e:
+            logger.error(f"Invalid YAML format: {e}")
+            raise EngineError(f"Invalid YAML format: {e}")
+
+    @classmethod
     def from_config(
         cls,
         config: AssistantConfig,
@@ -87,64 +111,10 @@ class ChatAssistantClient(BaseAssistantClient):
         """
         try:
             config_json = config.to_json()
-            is_create = not config.assistant_id
-            return cls(config_json=config_json, callbacks=callbacks, is_create=is_create, timeout=timeout)
+            return cls.from_json(config_json, callbacks, timeout)
         except Exception as e:
             logger.error(f"Failed to create chat client from config: {e}")
             raise EngineError(f"Failed to create chat client from config: {e}")
-
-    def _init_chat_assistant_client(
-            self, 
-            config_data: dict,
-            is_create: bool = True,
-            timeout: Optional[float] = None
-    ):
-        try:
-            # Create or update the assistant
-            assistant_config = AssistantConfig.from_dict(config_data)
-            if is_create:
-                assistant_config.assistant_id = str(uuid.uuid4())
-            self._reset_system_messages(assistant_config)
-            self._update_tools(assistant_config)
-            self._load_selected_functions(assistant_config)
-            self._assistant_config = assistant_config
-
-            # Update the local configuration using AssistantConfigManager
-            # TODO make optional to save the assistant_config in the config manager
-            config_manager = AssistantConfigManager.get_instance()
-            config_manager.update_config(self._name, assistant_config.to_json())
-
-        except Exception as e:
-            logger.error(f"Failed to initialize assistant instance: {e}")
-            raise EngineError(f"Failed to initialize assistant instance: {e}")
-
-    def purge(
-            self,
-            timeout: Optional[float] = None
-    )-> None:
-        """
-        Purges the chat assistant from the local configuration.
-
-        :param timeout: The HTTP request timeout in seconds.
-        :type timeout: Optional[float]
-
-        :return: None
-        :rtype: None
-        """
-        try:
-            logger.info(f"Purging chat assistant with name: {self.name}")
-            # retrieve the assistant configuration
-            config_manager = AssistantConfigManager.get_instance()
-            assistant_config = config_manager.get_config(self.name)
-
-            # remove from the local config
-            config_manager.delete_config(assistant_config.name)
-
-            self._clear_variables()
-
-        except Exception as e:
-            logger.error(f"Failed to purge chat assistant with name: {self.name}: {e}")
-            raise EngineError(f"Failed to purge chat assistant with name: {self.name}: {e}")
 
     def process_messages(
             self, 
@@ -258,10 +228,6 @@ class ChatAssistantClient(BaseAssistantClient):
             logger.error(f"Error occurred during processing run: {e}")
             raise EngineError(f"Error occurred during processing run: {e}")
 
-    def _reset_system_messages(self, assistant_config: AssistantConfig):
-        instructions = self._replace_file_references_with_content(assistant_config)
-        self._messages = [{"role": "system", "content": instructions}]
-
     def _handle_non_streaming_response(self, response, thread_name, run_id):
         response_message = response.choices[0].message
         self._messages.append(response_message)
@@ -314,16 +280,6 @@ class ChatAssistantClient(BaseAssistantClient):
 
         return tool_calls, collected_messages
 
-    def _append_tool_calls(self, tool_calls, tcchunklist):
-        for tcchunk in tcchunklist:
-            while len(tool_calls) <= tcchunk.index:
-                tool_calls.append({"id": "", "type": "function", "function": {"name": "", "arguments": ""}})
-            tc = tool_calls[tcchunk.index]
-            tc["id"] += tcchunk.id or ""
-            tc["function"]["name"] += tcchunk.function.name or ""
-            tc["function"]["arguments"] += tcchunk.function.arguments or ""
-        return tool_calls
-
     def _process_tool_calls(self, tool_calls, run_id):
         if tool_calls:
             logger.info(f"Tool calls: {tool_calls}")
@@ -362,19 +318,3 @@ class ChatAssistantClient(BaseAssistantClient):
                 metadata={"chat_assistant": self._name}
             )
             logger.info("Messages updated in conversation.")
-
-    def _update_tools(self, assistant_config: AssistantConfig):
-        logger.info(f"Updating tools for assistant: {assistant_config.name}")
-        if assistant_config.selected_functions:
-            self._tools = []
-            modified_functions = []
-            for function in assistant_config.selected_functions:
-                # Create a copy of the function spec to avoid modifying the original
-                modified_function = copy.deepcopy(function)
-                # Remove the module field from the function spec
-                if "function" in modified_function and "module" in modified_function["function"]:
-                    del modified_function["function"]["module"]
-                modified_functions.append(modified_function)
-            self._tools.extend(modified_functions)
-        else:
-            self._tools = None
